Don't show the prompt until the connection is established
72f81bed616aab12a5ab33167e470e51a7b359e9 changed things so that
the first call to Readline() happens immediately, even before
the connection is established. If the server is unavailable,
this leads to a confusing user experience (the prompt is displayed,
the first line of input is accepted, then the program hangs). Fix this
by pausing the Readline() loop until the connection is established.